+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Change log
+## 1.1-SNAPSHOT [Unreleased]
+### Added
+* Map-like extension functions for Table (#3)
+    * `forEach`
+    * `all` / `any` / `none`
+    * `count`
+    * `maxBy` / `maxWith` / `minBy` / `minWith`
+
 ## 1.0.0 (2017-01-14)
 ### Added
 * Interface Table